Projects/detect/requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
entrypoints==0.3
flake8==3.8.4
idna==2.6
importlib-metadata==3.3.0
keyring==17.1.1
keyrings.alt==3.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pycrypto==2.6.1
pyflakes==2.2.0
PyGObject==3.30.4
python-apt==1.8.4.1
pyxdg==0.25
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
ssh-import-id==5.7
typing-extensions==3.7.4.3
urllib3==1.24.1
zipp==3.4.0

